Adding a model in the project
Currently we have our model related code (which at this point is very simple in its scope) inside the products.js controller file.
In the products.js file, we have a 'products' array, and a single product is created on the fly as an object with a title, inside the 'postAddProduct' middleware function, after which we push that product into the products array. This all happens when the user enters a new product name in the input field of the browser in the /admin/add-product path. So the individual product created and the products array represents the data in our app (later we will have users as well).
So even though we aren't doing much with the product as of now, we can create a model and move all the data related code in that model.

* Create a new folder in the root directory called 'models' and inside that create a new file called 'product.js' (Its 'product' and not 'products').

Earlier we were creating a single product on the fly using the object literal syntax that only had a title. Now we will create a class which will be the blueprint of a product and we can create as many products as we want by calling the constructor with the 'new' keyword.
So the reason the we are creating a product class inside the model is because models represnt the data and business logic of our application and in this case, our core data is ultimately a product. Sure we also have lists of products with which we work, but the core thing that makes up the app is how a product looks like. What fields it has, does is have an image, title, etc. That is our core data. A list of products is essentially more of the same product, and therefore a single product is what defines our app in the end.

* So in product.js, create the same empty array we created in products.js called 'products'.

* Then create a 'Product' class and export it using 'module.exports'
In the class, setup a 'title' property using the constructor, create an instance method called 'save()' which when invoked, will push the object that is calling it, into the products array. And create a static 'fetchAll' method, that when invoked on the class, will retrieve all the products available in the products array.
The fetchAll() method is static because its not called on a single instance of the product
because it should fetch all products, and we don't want to create a new object with the 'new' keyword with some dummy title property just to fetch all existing products so therefore we want to call the fetchAll() static method directly on the class itself and not on an instantiated object. Inside the static method, we will simply return the 'products' array that we had defined at the top of the product.js page, above the class.

* Next go into the products.js controller file and first get rid of the products array as now we have it in the product.js model file,

* Then inside the postAddProduct middleware function we will remove the following line of code:
products.push({ title: req.body.title });
This is because now the product.js model file will be handling the pushing of the product into the products array, and also the creation of a new product instance.

So now we don't have any products array related logic in the products.js controller file any longer.

* In products.js controller file import the Product class

* Inside the postAddProduct function, above the res.redirect() function, add in the following code:

const product = new Product(req.body.title);

So we have created a new local constant called product and using the 'new' keyword, we instantiate a new Product instance and pass the 'req.body.title' value that we receive from the user input in the /add-product page as an argument. This argument will then create a new Product instance with a key of 'title' and a value of whatever the value of 'req.body.title' holds.

* Next under the newly instantiated product, inside the postAddProduct function, we want to save this newly created product instance into the products array and we do that with the following code:

product.save();

So this is what the newly updated postAddProduct function looks like in the products.js controller file:

exports.postAddProduct = (req, res, next) => {
  const product = new Product(req.body.title);
  product.save();
  res.redirect("/");
};

* Now we also want to get all the products in the array when we are on the '/' path and thats why inside the getProducts() middleware function in products.js, above the res.render() function we will add the following line of code:

const products = Product.fetchAll();

So the getProducts() middleware function will look like this:

exports.getProducts = (req, res, next) => {
  const products = Product.fetchAll();
  res.render("shop", {
    prods: products,
    pageTitle: "Shop",
    path: "/",
  });
};

When we use the static method fetchAll(), the invoking of that function will return to us the products array we defined in the product.js page (maybe with some items, or maybe empty) and that array will be saved into the constant 'products' that we defined in the getProducts() middleware function.